move nette files to bridge directory
