+++ src/Bridge/Nette/Response/PdfResponse.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace Contributte\Invoice\Responses;
+namespace Contributte\Invoice\Bridge\Nette\Response;
 
 use Nette\Application\Response;
 use Nette\Http\IRequest;
@@ -18,6 +18,7 @@ final class PdfResponse implements Response
 	public function send(IRequest $httpRequest, IResponse $httpResponse): void
 	{
 		$httpResponse->setContentType('application/pdf', 'utf-8');
+
 		echo $this->content;
 	}
